% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_list_artifacts}
\alias{gh_list_artifacts}
\title{List all artifacts of a repo
List all artifacts of a repo or single run}
\usage{
gh_list_artifacts(repo, runid = NULL, pat = gh_pat())
}
\description{
List all artifacts of a repo
List all artifacts of a repo or single run
}
