% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_list_run_jobs}
\alias{gh_list_run_jobs}
\title{Show all jobs of a run. Some functions require the corresponding job id.}
\usage{
gh_list_run_jobs(repo, runid, pat = gh_pat())
}
\description{
Show all jobs of a run. Some functions require the corresponding job id.
}
