% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_cli.R
\name{gh_new_repo}
\alias{gh_new_repo}
\title{Create a new repo on github and locally}
\usage{
gh_new_repo(
  reponame,
  parentdir,
  access = c("public", "private")[2],
  pat = gh_pat()
)
}
\description{
Create a new repo on github and locally
}
