% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_newest_runid}
\alias{gh_newest_runid}
\title{Get the id if the newest workflow run of a repo.}
\usage{
gh_newest_runid(repo, pat = gh_pat())
}
\description{
Get the id if the newest workflow run of a repo.
}
