% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{gh_remove_history}
\alias{gh_remove_history}
\title{Removes all old commits of a branch and only keeps current local files}
\usage{
gh_remove_history(repodir, branch = "main")
}
\description{
Based on https://superuser.com/questions/926915/how-can-i-delete-old-commits-in-github-via-terminal
WARNING: Don't use for real code repos, in particular not for repos multiple users work with.
}
