% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_run_workflow}
\alias{gh_run_workflow}
\title{Trigger the specified workflow}
\usage{
gh_run_workflow(
  repo,
  name = NULL,
  branch = "main",
  inputs = NULL,
  pat = gh_pat()
)
}
\arguments{
\item{repo}{The repository name in format user/reponame.}

\item{name}{The name of the workflow. If NULL use the first workflow}

\item{branch}{The Github branch. By default "main"}

\item{inputs}{Not yet tested. Potential input parameters passed to the workflow.}
}
\description{
Only works if the workflow YAML file sets
}
\details{
on:
  workflow_dispatch:
}
