% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_set_pat}
\alias{gh_set_pat}
\title{Set personal access token by writing it into environment variables.}
\usage{
gh_set_pat(pat)
}
\arguments{
\item{pat}{A string containing the personal access token generated on Github.

Not very secure. At some point I should use the approach of the gh package}
}
\description{
Set personal access token by writing it into environment variables.
}
