% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_cli.R
\name{gh_set_secret}
\alias{gh_set_secret}
\title{Set a repository secret using github client}
\usage{
gh_set_secret(repodir, name, value)
}
\arguments{
\item{repodir}{path to your local repository}

\item{name}{name of the secret}

\item{value}{of the secret}
}
\description{
Set a repository secret using github client
}
