% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{gh_update}
\alias{gh_update}
\title{Updates the Github repository to the new content of the local repo directory}
\usage{
gh_update(repodir, msg = "update", remote = "origin", branch = "main")
}
\description{
Updates the Github repository to the new content of the local repo directory
}
