% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_api.R
\name{gh_wait_until_new_run_starts}
\alias{gh_wait_until_new_run_starts}
\title{Wait until a new Github action job has started}
\usage{
gh_wait_until_new_run_starts(
  repo,
  old_runid,
  pause.sec = 1,
  timeout = 60,
  pat = gh_pat()
)
}
\description{
Wait until a new Github action job has started
}
